refactor: config.rs removed from the eth crate and LogStateUpdate moved to state_update.rs
