+++ crates/client/eth/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod utils;
-
 pub mod client;
-pub mod config;
 pub mod error;
 pub mod state_update;
